Add files via upload
"binary_search.py"  This thing does work... it's just that I made it to run with a stack in a separate file. I need to upload that. I've been working a lot on branching with git lately. Still, the concept is a little too abstract for me to apply it well enough yet. A good way to practice and to make this algorithm "go" would be to commit the stack and any files that refer to it in a separate branch. I hope I read this memo later.