Add files via upload
Uploading an unfinished (I literally have to fix one thing) Euclidian GCD program. Basically, what it is supposed to do is find the gcd of 'a' and 'b' by using recursion and a%b as the 'a' argument until there is no modulus and 'b' is returned as the GCD. My description may be missing a few details, but if you don't know what the Euclidian algorithm is then you don't have any business critiquing my efforts of explaining it to you. Who are you anyway? How did you even get here? 
Do you want to build an app together?